Add evaluateEquation function
Modify breakDownEquation for better error handling for invalid user
input like whitespace or a lone decimal. There remains a bug because of
how parseFloat is handled (e.g. parseFloat(2  xxx 111) returns 2), but
this should not matter as error checking is in place for everything the
user will be able to input in the final program